Initial Client Message Support
Migrated entire client-side message processing system to use the new auto-serializing messages.

Even more so than the server message support, this is currently almost certainly broken to some degree, and has not been tested.

Unlike the server messages, which were mostly organized in a manner that was easy to migrate, client message processing is quite spread out, and so supporting the new messages required some guesswork to properly implement.