Hotfix: Fix Missing `readMessageType()` Calls
At the moment, no checking is done on the result; it is assumed the server has sent the expected message.

In the future, the client should take a stance on either validating message types, or not reading them at all (this would require `Message` support to not write the value).