Mild Server Message Improvements
Made use of the new message features to provide typed message fields to server/UpdateMessage.

Most of the server messages do not need further modification as they are already natively strings. The only exception is server/FlipMessage and server/RollMessage, which will be updated in a future commit due to potentially requiring deep changes to the client.